Refactor search component to cancel on ESC key
