+++ src/lib/dataQuery.js
@@ -2,10 +2,16 @@ import fetch from 'isomorphic-fetch'
 
 const BUZZLER_BACKEND_URL = '/api/graphql'
 
-const dataFetch = async (query) => {
+export const dataFetch = async (query) => {
   const response = await fetch(`${BUZZLER_BACKEND_URL}?query=${query}`)
   const { data } = await response.json()
   return { ...data }
 }
 
-export default dataFetch
+export const dataMutation = async (query) => {
+  const response = await fetch(`${BUZZLER_BACKEND_URL}?query=${query}`, {
+    method: 'POST'
+  })
+  const { data } = await response.json()
+  return { ...data }
+}
